==> code/0_master.R <==
# ------------------------------------------------------------------------
# Master file: Projects BEV fleet (2050) 
#
# Julio Barros 
# 25/07/2023
# ------------------------------------------------------------------------


# packages ----------------------------------------------------------------

library(tidyverse)
library(readxl)
library(zoo)
library("ggsci")
library("ggthemes")


# run files ---------------------------------------------------------------

source('./code/1_anfavea_registration_yearly_files.R')
source('./code/2_anfavea_registration.R')
source('./code/3_compute_fleet.R')
source('./code/4_projection_inertial.R')
source('./code/5_projection_optimist.R')
source('./code/6_projection_biofuels.R')
source('./code/7_ancillary_graphs.R')


# finish ------------------------------------------------------------------

rm(list = ls())
gc()
